lib/weak-supervision/pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool]

[tool.black]
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
include = '\.pyi?$'
line-length = 100
target-version = ['py36', 'py37', 'py38', 'py39']

[tool.coverage.report]
exclude_lines = [
  "if __name__ == .__main__.:",
  "pass",
  "pragma: no cover",
  "raise NotImplementedError",
  "raise TypeError",
  "raise ValueError"
]

[tool.isort]
known_local_folder = [
  'tests',
  'weak-supervision'
]
line_length = 100
profile = 'black'

[tool.poetry]
authors = ["Arkhn's AI Team <ai@arkhn.com>"]
description = "A project to train and evaluate models with weak supervision techniques"
license = "Apache-2.0"
name = "weak-supervision"
readme = "README.md"
repository = "https://github.com/arkhn/ai-lembic/tree/main/experiements/weak-supervision"
version = "0.1.0"

[tool.poetry.dependencies]
bs4 = "^0.0.1"
datasets = "^2.8.0"
hydra-colorlog = "^1.2.0"
hydra-core = "^1.3.1"
hydra-joblib-launcher = "^1.2.0"
hydra-optuna-sweeper = "^1.2.0"
kmeans-pytorch = "^0.3"
lxml = "^4.9.2"
openai = "^0.26.2"
pandas = "^1.5.3"
protobuf = "3.20.1"
pyrootutils = "^1.0.4"
python = "~3.9"
pytorch-lightning = "^1.8.3"
requests = "^2.28.2"
rich = "^13.2.0"
scikit-learn = "^1.2.1"
sentence-transformers = "^2.2.2"
spacy = "^3.5.1"
syntok = "^1.4.4"
tenacity = "^8.1.0"
torch = "^1.13.1"
torchmetrics = "^0.11.0"
tqdm = "^4.64.1"
transformers = "^4.11.0"
typer = "^0.7.0"
wandb = "^0.14.2"

[tool.poetry.group.test.dependencies]
pytest = "^7.1.1"
pytest-cov = "^3.0.0"

[tool.pytest.ini_options]
addopts = "--cov-report term-missing:skip-covered"
markers = [
  "serial",
  "slow: marks tests as slow (deselect with '-m \"not slow\"')"
]
